### 2019-05-30 #1481 @whedon
[PRE REVIEW]: `hopit`: An R Package for Analysis of Reporting Behavior Using Generalized Ordered Probit Models
**Submitting author:** @MaciejDanko (<a href="http://orcid.org/0000-0002-7924-9022">Maciej J. Dańko</a>)
**Repository:** <a href="https://github.com/MaciejDanko/hopit" target ="_blank">https://github.com/MaciejDanko/hopit</a>
**Version:** 0.9.2.9000
