### 2019-05-30 #1481 @kyleniemeyer
Hi @MaciejDanko, our automated system is having some trouble compiling your paper—I think it is the (attempted) inclusion of the `float` package, and then using the LaTeX figure commands in the Markdown paper. Can you instead try to use Markdown figure commands?

@arfon, can you comment on whether this *should* be possible? I know we have been working on supporting fully LaTeX-based submissions, if that is necessary for what the author wants.
