### 2019-05-30 #1481 @arfon
> @arfon, can you comment on whether this should be possible? I know we have been working on supporting fully LaTeX-based submissions, if that is necessary for what the author wants.

If possible, we strongly prefer people to use standard Markdown syntax here for including figures.

e.g. `Figures can be included like this: ![Example figure.](figure.png)`
