### 2019-05-30 82d8058
LaTeX removed from paper.md
