+++ paper/paper.md
@@ -7,7 +7,6 @@ tags:
 - latent index
 - latent health
 - disability weights
-# author: Maciej J. Dańko
 authors:
 - name: Maciej J. Dańko
   affiliation: 1
@@ -18,11 +17,7 @@ affiliations:
 date: "3 May 2019"
 bibliography: paper.bib
 link-citations: true
-header-includes: \usepackage{float}
-# output:
-#   pdf_document:
-#     highlight: tango
-#     number_sections: no
+# header-includes: \usepackage{float}
 ---
 # Summary
 ``hopit`` is an open source software library written in the R [@Rteam2019] and C++ [@RcppEigen2013; @Rcpp2011] programming languages. The ``hopit`` package provides versatile methods for fitting and analyzing ordered response data in the context of heterogeneity in self reporting behavior.
@@ -37,25 +32,27 @@ The ``hopit`` package delivers functions and methods to fit (```hopit```), summa
 
 Once the model is fitted, the model estimates are used to determine reporting behavior; i.e., how the continuous latent measure is projected onto the categorical response. In practice, this is done by comparing actual categorical ordered responses with theoretical responses that are adjusted for heterogeneity in reporting behaviors, and are, therefore, more comparable across individuals.
 
+<!--
 \begin{figure}[H]
 {\centering\includegraphics{HI-2.pdf}
 
 }
 \caption[Health index vs. self-reported health for a model fitted to an exemplary data.]{Health index vs. self-reported health for a model fitted to an exemplary data.}\label{fig:HI-2}
-\end{figure}
-<!-- ![Health index vs. self-reported health for a model fitted to an exemplary data.](HI-2.pdf) -->
+\end{figure}-->
+![Health index vs. self-reported health for a model fitted to an exemplary data.](HI-2.pdf)
 
 One of the first steps of the analysis is the standardization of the latent measure to obtain the latent index. In the self-rated health example the latent health index is a proxy for the true underlying health of an individual, and varies from 0, representing the (model-based) worst health state in the sample, to 1, representing the (model-based) best health state in the sample (Fig. 1).
 
 The predicted latent measure obtained from the model is also used to standardize the latent variable coefficients. In the self-rated health example the standardized coefficients are called disability weights [@Jurges2007; @OKSUZYAN2019] and are calculated for each health variable to provide information about the impact of a specific health measure on the latent index (Fig. 2). The disability weight for a health variable is equal to the ratio of the corresponding health coefficient and the difference between the lowest and the highest values of predicted latent health. In other words, the disability weight reduces health index by some given amount or percentage; i.e. the health index of every individual is reduced by the same amount if the person had a heart attack or other heart problems)[@Jurges2007; @OKSUZYAN2019]. While the latent index is intended to reflect the underlying health, happiness or other status across individuals, the standardized coefficients (disability weights for health case), are computed for an average individual in the study population. 
 
+<!-- 
 \begin{figure}[H]
 {\centering\includegraphics{D-2.pdf}
 
 }
 \caption[Disability weights for the fitted model.]{Disability weights for the fitted model.}\label{fig:D-2}
-\end{figure}
-<!-- ![Disability weights for the fitted model.](D-2.pdf) -->
+\end{figure} -->
+![Disability weights for the fitted model.](D-2.pdf) 
 
[…]
