### 2019-05-31 bb9a696
LaTeX removed from the paper
