### 2019-05-31 #1481 @MaciejDanko
Thank you for your quick response,
paper.md is written in the standard Markdown syntax now
