### 2019-05-31 #1481 @MaciejDanko
Dear all,
Thank you for your quick response
I removed LaTeX style from md file.
The problem is that the last figure appears now after acknowledgements.
Best wishes,
Maciej
