### 2019-05-31 #1481 @MaciejDanko
I checked the paper proof and found that there is a problem with figure 3 caption.
The md code looks OK, I don't know why there is a problem.
