### 2019-06-08 #1481 @MaciejDanko
... I think this is Whedon's problem, I could generate the pdf without mistakes.
Anyway, I shortened the caption of the figure 3 and added additional blank lines
Could you check if it works now?
