### 2019-06-08 #1481 @arfon
@MaciejDanko - you had a typo in your `paper.md` which I've fixed in https://github.com/MaciejDanko/hopit/pull/5
