### 2019-06-08 #1481 @MaciejDanko
I don't know why I didn't notice the typos.
Thank you very much!
paper.md is updated now.
