### 2019-06-08 #1481 @MaciejDanko
Proof checked. No problems found. Thank you.
