### 2019-06-17 #1508 @whedon
[REVIEW]: `hopit`: An R Package for Analysis of Reporting Behavior Using Generalized Ordered Probit Models
**Submitting author:** @MaciejDanko (<a href="http://orcid.org/0000-0002-7924-9022">Maciej J. Dańko</a>)
**Repository:** <a href="https://github.com/MaciejDanko/hopit" target ="_blank">https://github.com/MaciejDanko/hopit</a>
**Version:** v0.10.1.9000

- [x] **Version:** v0.10.1.9000

- [x] **Version:** v0.10.1.9000
