### 2019-06-24 5d52426
JOSS review, edits
