+++ paper/paper.bib
@@ -7,6 +7,7 @@ volume = {9},
 number = {1},
 pages = {1-19},
 note = {R package verson 2.2},
+doi = "10.18637/jss.v009.i08",
 }
 
 @Book{rmarkdown2018,
@@ -62,7 +63,6 @@ note = {R package version 0.7},
 url = {https://CRAN.R-project.org/package=rticles},
 }
 
-
 @Misc{Lumley2019,
 author = {Thomas Lumley},
 year = {2019},
@@ -96,7 +96,6 @@ number="1",
 pages="1-11",
 issn="0361-0926",
 doi="10.1080/03610928508828893",
-url="https://doi.org/10.1080/03610928508828893"
 }
 
 @Article{Rebelo2014,
@@ -137,6 +136,7 @@ volume="4",
 number="1",
 journal="Journal of Mathematical Sociology",
 author="McKelvey, R. D. and Zavoina, W.",
+doi = "10.1080/0022250X.1975.9989847",
 year="1975",
 pages="103–120"
 }
@@ -213,7 +213,7 @@ volume = "228",
 pages = "41 - 50",
 year = "2019",
 issn = "0277-9536",
-doi = "https://doi.org/10.1016/j.socscimed.2019.03.002",
+doi = "10.1016/j.socscimed.2019.03.002",
 author = "Anna Oksuzyan and Maciej J. Dańko and Jennifer Caputo and Domantas Jasilionis and Vladimir M. Shkolnikov",
 }
 
@@ -232,6 +232,7 @@ url = "https://www.r-project.org"
     volume = "52",
     number = "5",
     pages = "1--24",
+    doi = "10.18637/jss.v052.i05",
     url = {http://www.jstatsoft.org/v52/i05/},
 }
 
@@ -244,7 +245,6 @@ url = "https://www.r-project.org"
     volume = "5",
     pages = "e3188v1",
     issn = "2167-9843",
-    url = "https://doi.org/10.7287/peerj.preprints.3188v1",
     doi = "10.7287/peerj.preprints.3188v1",
   }
 
@@ -256,6 +256,5 @@ url = "https://www.r-project.org"
     volume = "40",
     number = "8",
     pages = "1--18",
-    url = "http://www.jstatsoft.org/v40/i08/",
     doi = "10.18637/jss.v040.i08",
   }
