+++ README.md
@@ -244,7 +244,7 @@ sc <- standardizeCoef(model1, plotf = TRUE, namesf = txtfun,
                       mar = c(10, 4, 1, 1))
 ```
 
-<img src="man/figures/README-unnamed-chunk-6-1.png" width="100%" />
+<img src="man/figures/README-fig1-1.png" width="65%" style="display: block; margin: auto;" />
 
 **Figure 1.** Disability weights.
 
@@ -274,24 +274,26 @@ index (or the latent index). It is simply calculated using the
 TRUE`).
 
 ``` r
-hi <- latentIndex(model1, plotf = TRUE, response = "data", 
-                  ylab = 'Health index', col='deepskyblue3')
+pmar <- par('mar'); par(mar = c(5.5, 4, 0.5, 0.5))
+hi <- latentIndex(model1, plotf = TRUE, response = "data", xlab='',
+                  ylab = 'Health index', col='deepskyblue3'); par(mar = pmar)
 ```
 
-<img src="man/figures/README-unnamed-chunk-8-1.png" width="100%" />
+<img src="man/figures/README-fig2-1.png" width="65%" style="display: block; margin: auto;" />
 
 **Figure 2.** Health index vs. self-reported health.
 
 The main aim of the reporting heterogeneity analyses is to determine the
 cut-points used to calculate the adjusted health status for each
 individual. The calculation and the plotting of cut-points are realized
-using the `getCutPoints` function.
+using the `getCutPoints`
+function.
 
 ``` r
 z <- getCutPoints(model = model1, decreasing.levels = TRUE)
 ```
 
-<img src="man/figures/README-unnamed-chunk-9-1.png" width="100%" />
+<img src="man/figures/README-fig3-1.png" width="65%" style="display: block; margin: auto;" />
 
 **Figure 3.** Health index cut-points for adjusted health levels.
 
@@ -322,7 +324,7 @@ hl <- getLevels(model = model1, formula = ~ sex + ageclass, data = healthsurvey,
                       sep = ' ', plotf = TRUE)
 ```
 
-<img src="man/figures/README-unnamed-chunk-13-1.png" width="100%" />
+<img src="man/figures/README-fig4-1.png" width="73%" style="display: block; margin: auto;" />
 
 **Figure 4.** Adjusted and original health levels.
 
@@ -366,7 +368,7 @@ calculate the mentioned difference:
 
 ``` r
 diff_BadHealth <- function(model, data) {
-  hl <- getLevels(model = model1, formula = ~ sex + ageclass, data = data, 
+  hl <- getLevels(model = model, formula = ~ sex + ageclass, data = data, 
                   sep = ' ', plotf = FALSE)
   hl$original[,1] + hl$original[,2] - hl$adjusted[,1]- hl$adjusted[,2]
 }
@@ -397,17 +399,17 @@ The code below plots the differences and the bootstrapped confidence
 intervals
 
 ``` r
-pmar <- par('mar'); par(mar = c(9.5, 6, pmar[3:4]))
+pmar <- par('mar'); par(mar = c(7, 5, 0.5, 0.5))
 m <- max(abs(est.CI))
 pos <- barplot(est.org, names.arg = names(est.org), las = 3, 
- ylab = 'Original - adjusted differences \nin the prevalences of bad health (%)', 
- ylim = c(-m, m), density = 20, angle = c(45, -45), col = c('blue', 'orange'))
+ylab = 'Original - adjusted differences \nin the prevalences of bad health (%)', 
+ylim = c(-m, m), density = 20, angle = c(45, -45), col = c('blue', 'orange'))
 for (k in seq_along(pos)) 
-  lines(c(pos[k,1],pos[k,1]), est.CI[,k], lwd = 2, col = 2)
[…]
