+++ README.md
@@ -99,25 +99,25 @@ for technical details about the SHARE database. The structure of
 ``` r
 data("healthsurvey")
 print(t(healthsurvey[1:6,-1]), quote=FALSE, na.print='NA', right=TRUE)
->                                1       2       3       4         5       6
-> health                 Very good    Good    Good    Good Excellent    Good
-> diabetes                      no      no     yes      no        no      no
-> obese                         no      no      no      no        no      no
-> IADL_problems                 no      no      no      no        no      no
-> hypertension                  no     yes      no      no        no     yes
-> high_cholesterol              no     yes      no      no        no     yes
-> respiratory_problems          no      no      no      no        no     yes
-> heart_attack_or_stroke        no     yes      no      no        no      no
-> poor_mobility                 no      no      no      no        no     yes
-> very_poor_grip                no      no      no      no        no      no
-> depression                    no      no      no     yes        no      no
-> other_diseases               yes     yes      no      no        no     yes
-> sex                          man     man     man     man     woman     man
-> ageclass                     80+   70-79   50-59   60-69       80+     80+
-> education                  prim-   prim-   prim-    sec+     prim-   prim-
-> country                        Y       Y       X       Y         Z       Y
-> csw                      2407.48 1198.12  885.26  772.04   1304.24  917.16
-> psu                           YB      YB      XC      YA        ZB      YD
+##                                 1       2       3       4         5       6
+##  health                 Very good    Good    Good    Good Excellent    Good
+##  diabetes                      no      no     yes      no        no      no
+##  obese                         no      no      no      no        no      no
+##  IADL_problems                 no      no      no      no        no      no
+##  hypertension                  no     yes      no      no        no     yes
+##  high_cholesterol              no     yes      no      no        no     yes
+##  respiratory_problems          no      no      no      no        no     yes
+##  heart_attack_or_stroke        no     yes      no      no        no      no
+##  poor_mobility                 no      no      no      no        no     yes
+##  very_poor_grip                no      no      no      no        no      no
+##  depression                    no      no      no     yes        no      no
+##  other_diseases               yes     yes      no      no        no     yes
+##  sex                          man     man     man     man     woman     man
+##  ageclass                     80+   70-79   50-59   60-69       80+     80+
+##  education                  prim-   prim-   prim-    sec+     prim-   prim-
+##  country                        Y       Y       X       Y         Z       Y
+##  csw                      2407.48 1198.12  885.26  772.04   1304.24  917.16
+##  psu                           YB      YB      XC      YA        ZB      YD
 ```
 
[…]
