### 2019-06-24 93c76d5
remove math symbols from README.md
