+++ R/generics.R
@@ -120,18 +120,21 @@ print.vcov.hopit <- function(x, digits = 3L, ...){
 #' @export
 #' @author Maciej J. Danko
 #' @useDynLib hopit
-#' @usage \method{summary}{hopit}(object, robust.se = TRUE, ...)
+#' @usage \method{summary}{hopit}(object, robust.se, ...)
 #' @method summary hopit
 #' @keywords internal
 #' @importFrom Rcpp evalCpp
-summary.hopit <- function(object, robust.se = TRUE, ...){
+summary.hopit <- function(object, robust.se, ...){
+  if (missing(robust.se)) {
+    if (length(object$design)) robust.se <- FALSE else robust.se <- TRUE
+  }
   varcov <- vcov.hopit(object, robust.se, ...)
   robust.se <- attr(varcov, 'robust.vcov')
   dvcov <- diag(varcov)
   if (any(dvcov<0)) warning(hopit_msg(43),call.=NA)
   SE <- suppressWarnings(sqrt(abs(dvcov)))
   if (length(object$design)){
-    cat(hopit_msg(44))
+    message(hopit_msg(44))
   }
   if ((!robust.se) && (any(is.na(SE))) && !(length(object$design)))
     warning(call. = FALSE, hopit_msg(45))
@@ -176,6 +179,230 @@ print.summary.hopit <- function(x, ...){
   invisible(NULL)
 }
 
+#' @keywords internal
+#' @export
+#' @method print hopitDW
+#' @noRd
+print.hopitDW <- function(x, digits=4, quote=FALSE, ...){
+  x1 <- as.matrix(format(round(as.vector(x),digits),nsmall=digits, digits=digits,
+         scientific = FALSE))
+  dimnames(x1) <- dimnames(x)
+  print.default(x1, quote=quote, digits=digits, ...)
+}
+
+#' @keywords internal
+#' @export
+#' @method summary hopitDW
+#' @noRd
+summary.hopitDW<-function(object, ...){
+  x <- data.frame(cbind('Std. coef'=as.vector(object), attr(object,'Ptab')),
+                  check.names = FALSE)
+  class(x) <- c('summary.hopitDW','data.frame')
+  attr(x,'legend') <- attr(object,'legend')
+  return(x)
+}
+
+#' @keywords internal
+#' @export
+#' @method print summary.hopitDW
+#' @noRd
+print.summary.hopitDW <- function(x, digits=4, show.coef.names=TRUE, ...){
+  if (show.coef.names) {
+    g <- data.frame('Coefficient name' = x[,2],
+                  'Standardized coeficient'=
+                    format(round(x[,1],digits),nsmall=digits, digits=digits,
+                           scientific = FALSE),
+                  'Pr(>|z|)' = format(round(x[,3],digits),
+                         nsmall=digits, digits=digits, scientific = FALSE),
+                  x[,4],
+                  fix.empty.names = FALSE, check.names = FALSE)
+  } else {
+    g <- data.frame('Standardized coeficient'=
+                      format(round(x[,1],digits), nsmall=digits, digits=digits,
+                             scientific = FALSE),
+                    'Pr(>|z|)' = format(round(x[,3],digits),
+                          nsmall=digits, digits=digits, scientific = FALSE),
+                    x[,4],
+                    fix.empty.names = FALSE, check.names = FALSE)
+  }
+  rownames(g)<-rownames(x)
+  print(g)
+  cat("---\nSignif. codes: ",attr(x,'legend'),'\n')
+}
+
+#' Plotting standardized coefficients
+#' @keywords internal
+#' @export
+#' @method plot hopitDW
+#' @param x a object generated by \code{\link{standardizeCoef}} function.
+#' @param ordered a logical indicating whether to sort the disability weights.
+#' @param show.signif show significance codes.
[…]
