+++ R/hopit.R
@@ -501,6 +501,7 @@ hopit.control<-function(grad.eps = 3e-5,
 #' values from the \code{data}. Using \code{\link[stats]{na.pass}} will lead to an error.
 #' @importFrom stats na.fail
 #' @importFrom stats sigma
+#' @importFrom stats model.frame
 #' @importFrom stats coef
 #' @importFrom Rdpack reprompt
 #' @return a \code{hopit} object used by other functions and methods. The object is a list with the following components:
@@ -509,6 +510,7 @@ hopit.control<-function(grad.eps = 3e-5,
 #'  \item{hasdisp}{ a logical indicating whether fit.sigma was modeled.}
 #'  \item{use.weights}{ a logical indicating whether any weights were used.}
 #'  \item{weights}{ a vector with model weights.}
+#'  \item{frame}{ a model frame.}
 #'  \item{latent.formula}{ a latent formula used to fit the model.}
 #'  \item{latent.mm}{ a latent model matrix.}
 #'  \item{latent.terms}{ latent variables used, and their interactions.}
@@ -775,6 +777,7 @@ hopit<- function(latent.formula,
   if (missing(decreasing.levels)) decreasing.levels = NULL
   check_decreasing.levels(decreasing.levels, levels(model$y_i))
   if (!decreasing.levels) model$y_i <- factor(model$y_i, rev(levels(model$y_i)))
+  model$decreasing.levels <- decreasing.levels
   model$y_latent_i <- NA # latent
   model$Ey_i <- NA # ordinal classified utput
   model$J <- length(levels(model$y_i))
@@ -817,6 +820,9 @@ hopit<- function(latent.formula,
   model$weights <- as.vector(matrix(model$weights, 1L, model$N))
   #scaling weights
   model$weights <- model$N * model$weights / sum(model$weights)
+
+  model$frame <- cbind.data.frame(stats::model.frame(latent.formula, data),
+                                  stats::model.frame(thresh.formula, data))
 
   #calculate special matrices for gradient calaculation
   model <- calcYYY(model)
