### 2019-07-06 #1508 @MaciejDanko
@soodoku, I noticed that you haven't checked "functionality" in the functionality checklist.

I decided to improved the functionality of the package by adding generic functions (_print_, _summary_, _plot_) to _latentIndex_, _standardizeCoef_, _getLevels_, and _getCutPoints_. I also improved the code so the functions are easier to run (less 'initial' parameters) for a user. The vignette and readme were slightly modified to follow the code update.

The community guidelines are now both in the readme and the contributing files

I also put DOI everywhere where it was passible.
