### 2019-07-07 #1508 @soodoku
thanks @MaciejDanko. cool improvements. hope to test it out v. soon and update the checklist!
