### 2019-07-13 #1508 @soodoku
@MaciejDanko: went through the package and the health example to check functionality. things seem to pan out. my only rec. would be = to prove to yourself that you are doing the right thing computationally, use either an off-the-shelf package or a toy example that you can reason through. otherwise, god speed!
