### 2019-08-14 #1508 @MaciejDanko
@soodoku thank you very much for the review. I have already checked the package by the replicating the results of Jurges (2007, doi: 10.1002/hec.1134) model. While the data and the model slightly differed I got almost identical results.
