### 2019-09-01 #1508 @alexhanna
Messaged @mkearney to check if he was able to still serve as a reviewer on the package.
