### 2019-09-01 #1508 @alexhanna
The paper draft looks good to me. Just a reminder to update the ??? parts of the paper.md file once the review is completed. I'll give @mkearney two weeks to reply to the review.

@openjournals/jose-eics if the reviewer doesn't respond, should we try to find another review for this package before continuing on it?
