### 2019-09-02 #1508 @arfon
> @openjournals/jose-eics if the reviewer doesn't respond, should we try to find another review for this package before continuing on it?

@alexhanna - there are a couple of options here:

- Find another reviewer.
- If you're comfortable, you could do a quick review yourself.
