### 2019-09-03 #1508 @mkearney
My apologies for the delay. The end of summer/beginning of Fall got away from me!

I got the package installed, working, and everything looks good, @MaciejDanko and @alexhanna !!

A few minor notes:

1. I would agree with @soodoku that it'd be nice to see/compare similar estimates from libraries.
1. Overall test coverage is solid, but I would point out there are no tests for functions in [R/boot.R](https://github.com/MaciejDanko/hopit/blob/master/R/boot.R). The main functionality seems to be supported with automated tests.
1. The installation was a bit picky on a linux and a mac to me. I had to temporarily rename my `.Rprofile` to get it installed properly. SO this was 100% user error, but I don't usually have to rename my `.Rprofile` to get packages working properly.
1. The vignette(s) are only in the form of PDFs. Can source code be supplied? Or a script-only version of the README (which I found to be useful)?
