### 2019-09-04 #1508 @MaciejDanko
@mkearney, thank you for your review!

1. @mkearney and @soodoku, here is the mentioned "off-the-shelf-package-example" using a very simple model. I use the ``vglm`` function of the VGAM package. The ``vglm()`` and the ``hopit()`` fit different models, but for this particular (simple) example (and data ?) one can compare (at least) the fit of the health variable coefficients.

``library(VGAM)``
``library(hopit)``

``# Let us define a very simple model such that both packages can easily run it; i.e. ``
``# no interactions between threshold and health variables, ``
``# no survey design, no fitting of sigma, etc.``

``latent.formula <- health ~ hypertension + high_cholesterol +``
``    heart_attack_or_stroke + poor_mobility + very_poor_grip +``
``    depression + respiratory_problems +``
``    IADL_problems + obese + diabetes + other_diseases``

``thresh.formula <- ~ sex + ageclass + country``

``# Fit the hopit model``
``model1.hopit <- hopit(latent.formula = latent.formula,``
``                thresh.formula = thresh.formula,``
``                decreasing.levels = TRUE,``
``                data = healthsurvey)``

``# Fit the vglm model``
``healthsurvey$health <- ordered(healthsurvey$health)``
``thresh.right <- deparse(thresh.formula[[2]])``
``big.formula <- update.formula(latent.formula,paste('. ~ . +',thresh.right))``
``parallel.formula <- formula(paste('FALSE ~', thresh.right))``

``model1.vglm<-vglm(big.formula, data = healthsurvey,``
``            family = cumulative(parallel = parallel.formula, link = 'probit'))``

``# Can compare only health variables; thresholds are fitted differently``
``round(cbind(VGAM= - coef(model1.vglm)[5:15],``
``      hopit=coef(model1.hopit,aslist=TRUE)$latent.params),4)``

2. The bootstrapping (``boot_hopit``) is not present in automated test, because running it takes some time. The ``boot_hopit`` is also quite simple function, which is not so prone to errors or changes in the main algorithm. Nonetheless, all examples of the ``boot_hopit`` and vignette were tested and work fine.

3. The hopit package (installation, auto-tests, examples, etc.) was tested on Travis (Linux) and AppVeyor (Windows). It was also checked during submission to CRAN. I will try to figure out the issue with ``.Rprofile``, but so far I cannot replicate it. I cannot even guess why it could happen.

4. The script version of README can be seen on GitHub (click raw option or use this link [README](https://raw.githubusercontent.com/MaciejDanko/hopit/master/README.md)).  I am not sure if it is a common practice to supply a code for vignette on GitHub. Anyway, I am attaching the most recent version of the code here [hopit_vignette.zip](https://github.com/openjournals/joss-reviews/files/3573793/hopit_vignette.zip). I hope it will be helpful :).
