+++ README.md
@@ -1,9 +1,12 @@
 
-[![Travis-CI Build Status](https://travis-ci.org/MaciejDanko/hopit.svg?branch=master)](https://travis-ci.org/MaciejDanko/hopit)
-[![AppVeyor Build Status](https://ci.appveyor.com/api/projects/status/github/MaciejDanko/hopit?branch=master&svg=true)](https://ci.appveyor.com/project/MaciejDanko/hopit)
+[![CRAN\_Version](https://www.r-pkg.org/badges/version/hopit)](https://cran.r-project.org/package=hopit)
+[![Travis-CI Build
+Status](https://travis-ci.org/MaciejDanko/hopit.svg?branch=master)](https://travis-ci.org/MaciejDanko/hopit)
+[![AppVeyor Build
+Status](https://ci.appveyor.com/api/projects/status/github/MaciejDanko/hopit?branch=master&svg=true)](https://ci.appveyor.com/project/MaciejDanko/hopit)
 [![status](http://joss.theoj.org/papers/73b926670df79a6cfa48cffa7d4775a6/status.svg)](http://joss.theoj.org/papers/73b926670df79a6cfa48cffa7d4775a6)
-[![CRAN Badge1](https://cranlogs.r-pkg.org/badges/grand-total/hopit)](https://CRAN.R-project.org/package=hopit)
-[![CRAN Badge2](https://cranlogs.r-pkg.org/badges/hopit)](https://CRAN.R-project.org/package=hopit)
+[![CRAN\_Download\_Badge1](https://cranlogs.r-pkg.org/badges/grand-total/hopit)](https://CRAN.R-project.org/package=hopit)
+[![CRAN\_Download\_Badge2](https://cranlogs.r-pkg.org/badges/hopit)](https://CRAN.R-project.org/package=hopit)
 [![license](https://img.shields.io/badge/License-GPL%20v3-blue.svg)](https://github.com/MaciejDanko/hopit/blob/master/LICENSE)
 
 # R-package *hopit*: Hierarchical ordered probit models with application to reporting heterogeneity.
@@ -14,12 +17,12 @@ sources. For example, the evaluation of an individual’s own health may
 depend on previous medical diagnoses, functional status, and symptoms
 and signs of illness; as on well as life-style behaviors, including
 contextual social, gender, age-specific, linguistic and other cultural
-factors (Oksuzyan et al. 2019). The **hopit** package offers versatile
+factors (Oksuzyan et al. 2019). The **hopit** package offers versatile
 functions for analyzing different self-reported ordinal variables, and
 for helping to estimate their biases. Specifically, the package provides
 the function to fit a generalized ordered probit model that regresses
 original self-reported status measures on two sets of independent
-variables (King et al. 2004; Jurges 2007; Oksuzyan et al. 2019). The
+variables (King et al. 2004; Jurges 2007; Oksuzyan et al. 2019). The
 first set of variables (e.g., health variables) included in the
 regression are individual statuses and characteristics that are directly
 related to the self-reported variable. In the case of self-reported
@@ -29,7 +32,7 @@ anthropometric measures, and lifestyle behaviors. The second set of
 independent variables (threshold variables) is used to model cut-points
 between adjacent self-reported response categories as functions of
 individual characteristics, such as gender, age group, education, and
-country (Oksuzyan et al. 2019). The model helps to adjust for specific
+country (Oksuzyan et al. 2019). The model helps to adjust for specific
 socio-demographic and cultural differences in how the continuous latent
 health is projected onto the ordinal self-rated measure. The fitted
 model can be used to calculate an individual predicted latent status
[…]
