### 2019-09-04 e7d1f45
README and paper
