### 2019-09-24 #1508 @MaciejDanko
@mkearney,
I noticed emoji under my reply, but you haven't ticked "performance" on the functionality checklist off.
Please let me known if you have any questions. I would be happy to help
