### 2019-09-25 #1508 @alexhanna
👋 hi @mkearney can you check the performance button?
