### 2019-10-15 #1508 @mkearney
> @mkearney,
> I noticed emoji under my reply, but you haven't ticked "performance" on the functionality checklist off.
> Please let me known if you have any questions. I would be happy to help

I thought I fixed this two different times. That'll teach me to do Github things on my phone... 😅
