### 2019-10-22 #1508 @MaciejDanko
Hi @alexhanna, the performance box has been ticked, could we proceed further? :)
