### 2019-10-22 #1508 @alexhanna
👋 @MaciejDanko - the last step we need you to do is get a DOI number from Zenodo and post it here. We can then assign the repository a DOI and pass it off to the EIC for acceptance.
