### 2019-11-06 #1508 @MaciejDanko
@alexhanna - there is a DOI in article proof (10.21105/joss.01508) shall I use this one when uploading the package to Zenodo?
