### 2019-11-14 #1508 @alexhanna
Hi @MaciejDanko, I made a small PR for a missing DOI (https://github.com/MaciejDanko/hopit/pull/12). Once you accept this, we should be good to go.
