### 2019-11-14 #1508 @alexhanna
@MaciejDanko -- hrm, it looks like it didn't update. if you're not citing the book, you could also choose to remove it.
