### 2019-11-14 1304662
remove knitr
