+++ paper/paper.bib
@@ -36,18 +36,6 @@ note = {R package version 0.6.3},
 url = {https://CRAN.R-project.org/package=pander},
 }
 
-@Book{knitr2015,
-title = {Dynamic Documents with {R} and knitr},
-author = {Yihui Xie},
-publisher = {Chapman and Hall/CRC},
-address = {Boca Raton, Florida},
-year = {2015},
-edition = {2nd},
-note = {ISBN 978-1498716963},
-url = {https://yihui.name/knitr/},
-doi = {10.1201/b15166}
-}
-
 @Manual{rticles2019,
 title = {rticles: Article Formats for R Markdown},
 author = "JJ Allaire and Yihui Xie and {R Foundation} and Hadley Wickham and {Journal of Statistical Software} and Ramnath Vaidyanathan and {Association for Computing Machinery} and Carl Boettiger and {Elsevier} and Karl Broman and Kirill Mueller and Bastiaan Quast and Randall Pruim and Ben Marwick and Charlotte Wickham and Oliver Keyes and Miao Yu and Daniel Emaasit and Thierry Onkelinx and Alessandro Gasparini and Marc-Andre Desautels and Dominik Leutnant and {MDPI} and Oğuzhan Öğreden and Dalton Hance and Daniel Nüst and Petter Uvesten and Elio Campitelli and John Muschelli",
