### 2019-11-14 #1508 @alexhanna
Excellent. @openjournals/joss-eics, this one is also ready to go.
