### 2019-11-15 #1508 @labarba
Hi @MaciejDanko — I found no [releases](https://github.com/MaciejDanko/hopit/releases) in the repo. Please make a tagged release with the reviewed version of the software and post the version number here to update the metadata.
