### 2019-11-16 #1508 @labarba
Could you merge my small fix in the `.bib`?
