### 2019-11-17 #1508 @MaciejDanko
Hi @labarba - small fix was merged, thank you.
